Windows/requirements.txt
# ============================================
# NDHSM Windows 管理工具 - Python 依赖
# 相关文件: main.py, config.py, muip_client.py,
#           main_window.py, auto_setup.py, proxy_manager.py
# ============================================

# GUI 框架
PySide6>=6.6.0
PySide6-Fluent-Widgets>=1.5.0

# HTTP 请求
requests>=2.31.0

# RSA 加密 (MUIP API 认证)
pycryptodome>=3.20.0

# Git 操作
GitPython>=3.1.40
